Move music context into own file
